feat: ship RDS departures on DUPs
* Merge relevant `dup_new` candidate generators into `dup` and remove
  the `new_departures` variant.

* Remove the `VariantCanary` plug for now, since it's now unused.

* Update Inspector UI to not display the variant switcher option if no
  screen types have variants.

* Remove `Alert.fetch_or_empty_list` + `Vehicle.by_route_and_direction`,
  which were only used in the old departures generator.

* Make `CandidateGenerator.Widgets.Departures.fetch_section_departures`
  private and adjust a few tests accordingly, since the old departures
  generator was the only external caller.

* Remove now-unused DUP "headway mode" configuration keys.

* Remove special case in departure widget serialization that used a
  `Schedule` with `nil` arrival and departure times as a signal to
  produce an "overnight row". This signal was only used by the old
  generator; the new one uses a distinct "special trip type" flag.

* Remove now-unused `Headways.get_with_route/2` and its associated data.
