+++ lib/screens/headways.ex
@@ -4,7 +4,6 @@ defmodule Screens.Headways do
   when no live departures are available ("Trains every X-Y minutes").
   """
 
-  alias Screens.Routes.Route
   alias Screens.SignsUiConfig.Cache, as: SignsUi
   alias Screens.Stops.Stop
   alias Screens.Util
@@ -38,98 +37,6 @@ defmodule Screens.Headways do
     silver_chelsea: [7096..7097, 74_611..74_617, 74_630..74_637]
   }
 
-  # Mapping of parent station IDs and route IDs to headway keys. For brevity, omits the "place-"
-  # prefix which is currently common to all parent station IDs.
-  @stations %{
-    blue_trunk: {
-      ~w[Blue],
-      ~w[wondl rbmnl bmmnl sdmnl orhte wimnl aport mvbcl aqucl state gover bomnl]
-    },
-    glx_medford: {~w[Green-E], ~w[mdftf balsq mgngl gilmn esomr]},
-    glx_union: {~w[Green-D], ~w[unsqu]},
-    green_b: {
-      ~w[Green-B],
-      ~w[
-        bland
-        buest
-        bucen
-        amory
-        babck
-        brico
-        harvd
-        grigg
-        alsgr
-        wrnst
-        wascm
-        sthld
-        chswk
-        chill
-        sougr
-        lake
-      ]
-    },
-    green_c: {
-      ~w[Green-C],
-      ~w[smary hwsst kntst stpul cool sumav bndhl fbkst bcnwa tapst denrd engav clmnl]
-    },
-    green_d: {
-      ~w[Green-D],
-      ~w[fenwy longw bvmnl brkhl bcnfd rsmnl chhil newto newtn eliot waban woodl river]
-    },
-    green_e: {
-      ~w[Green-E],
-      ~w[prmnl symcl nuniv mfa lngmd brmnl fenwd mispk rvrwy bckhl hsmnl]
-    },
-    # Not all of these stations are served by all Green Line branches, but "trunk" is the most
-    # granular division of the Green Line we have to work with
-    green_trunk: {
-      ~w[Green-B Green-C Green-D Green-E],
-      ~w[lech spmnl north haecl gover pktrm boyls armnl coecl hymnl kencl]
-    },
-    orange_trunk: {
-      ~w[Orange],
-      ~w[
-        ogmnl
-        mlmnl
-        welln
-        astao
-        sull
-        ccmnl
-        north
-        haecl
-        state
-        dwnxg
-        chncl
-        tumnl
-        bbsta
-        masta
-        rugg
-        rcmnl
-        jaksn
-        sbmnl
-        grnst
-        forhl
-      ]
-    },
-    red_trunk: {
-      ~w[Red],
-      ~w[alfcl davis portr harsq cntsq knncl chmnl pktrm dwnxg sstat brdwy andrw jfk]
-    },
-    red_ashmont: {~w[Red], ~w[shmnl fldcr smmnl asmnl]},
-    red_braintree: {~w[Red], ~w[nqncy wlsta qnctr qamnl brntn]},
-    # combining 743 with the other three for the common stops
-    silver_seaport: {~w[741 742 743 746], ~w[conrd wtcst crtst sstat]},
-    silver_chelsea: {~w[743], ~w[aport estav boxdt belsq chels]}
-  }
-
-  # Mapping of stops and route IDs to headway keys for the Silver Line,
-  # for stops which serve more than one route.
-  @sl_stops %{
-    # congress_st_at_wtc 17_096
-    # combining 743 with the other three for the common stops
-    silver_seaport: {~w[741 742 743 746], ~w[17096]}
-  }
-
   @type range :: {low :: pos_integer(), high :: pos_integer()}
 
   @doc """
@@ -145,42 +52,12 @@ defmodule Screens.Headways do
     end
   end
 
-  @doc """
-  Gets headway values for a stop, with a route provided to disambiguate when the stop is a parent
-  station served by multiple routes.
-
-  ⚠️ Included for compatibility with existing DUP departures logic. Prefer `get/2`.
-  """
-  @callback get_with_route(Stop.id(), Route.id()) :: range() | nil
-  @callback get_with_route(Stop.id(), Route.id(), DateTime.t()) :: range() | nil
-  def get_with_route(stop_id, route_id, at \\ DateTime.utc_now()) do
-    case headway_key(stop_id, route_id) do
-      nil -> nil
-      key -> key |> SignsUi.headways() |> Map.get(period(at))
-    end
-  end
-
   @spec headway_key(Stop.id()) :: SignsUi.headway_key() | nil
-  @spec headway_key(Stop.id(), Route.id()) :: SignsUi.headway_key() | nil
   for {key, ranges} <- @stops, range <- ranges, stop_id <- range do
     defp headway_key(unquote(to_string(stop_id))), do: unquote(to_string(key))
-    defp headway_key(unquote(to_string(stop_id)), _route_id), do: unquote(to_string(key))
-  end
-
-  for {key, {route_ids, stations}} <- @stations,
-      route_id <- route_ids,
-      station <- stations do
-    defp headway_key("place-" <> unquote(station), unquote(route_id)), do: unquote(to_string(key))
-  end
-
-  for {key, {route_ids, stations}} <- @sl_stops,
-      route_id <- route_ids,
-      station <- stations do
-    defp headway_key(unquote(station), unquote(route_id)), do: unquote(to_string(key))
   end
 
   defp headway_key(_stop_id), do: nil
-  defp headway_key(_stop_id, _route_id), do: nil
 
   @spec period(DateTime.t()) :: :peak | :off_peak | :saturday | :sunday
   defp period(datetime) do
