+++ lib/screens/v2/candidate_generator/dup/departures.ex
@@ -1,793 +1,468 @@
 defmodule Screens.V2.CandidateGenerator.Dup.Departures do
   @moduledoc false
 
-  require Logger
+  import Screens.Inject
 
-  alias Screens.Alerts.{Alert, InformedEntity}
-  alias Screens.Report
+  alias Screens.Lines.Line
   alias Screens.Routes.Route
-  alias Screens.Schedules.Schedule
-  alias Screens.Stops.Stop
-  alias Screens.Util
-  alias Screens.V2.CandidateGenerator.Widgets
+
   alias Screens.V2.Departure
+  alias Screens.V2.RDS
+  alias Screens.V2.RDS.{Countdowns, FirstTrip, Headways, NoService, ServiceEnded}
+
   alias Screens.V2.WidgetInstance.Departures, as: DeparturesWidget
-  alias Screens.V2.WidgetInstance.{DeparturesNoData, OvernightDepartures}
-  alias Screens.Vehicles.Vehicle
-  alias ScreensConfig.{Departures, Screen}
-  alias ScreensConfig.Departures.{Header, Layout, Query, Section}
-  alias ScreensConfig.Departures.Query.Params
-  alias ScreensConfig.Screen.Dup
 
   alias Screens.V2.WidgetInstance.Departures.{
+    HeadwayRow,
     HeadwaySection,
     NoDataSection,
     NormalSection,
+    NoServiceSection,
     OvernightSection
   }
 
-  import Screens.Inject
+  alias Screens.V2.WidgetInstance.{DeparturesNoData, DeparturesNoService, OvernightDepartures}
 
-  @headways injected(Screens.Headways)
+  alias ScreensConfig.Departures
+  alias ScreensConfig.Departures.{Header, Layout, Section}
+  alias ScreensConfig.Screen
+  alias ScreensConfig.Screen.Dup
 
-  def departures_instances(
+  @type widget :: DeparturesNoData.t() | DeparturesWidget.t() | OvernightDepartures.t()
+  @rds injected(RDS)
+
+  @max_departures_per_rotation 4
+
+  @primary_slot_names [
+    :main_content_zero,
+    :main_content_one,
+    :main_content_reduced_zero,
+    :main_content_reduced_one
+  ]
+  @secondary_slot_names [:main_content_two, :main_content_reduced_two]
+
+  @spec instances(Screen.t(), DateTime.t()) :: [widget()]
+  def instances(
         %Screen{
           app_params: %Dup{
-            primary_departures: %Departures{sections: primary_sections},
-            secondary_departures: %Departures{sections: secondary_sections}
+            primary_departures: primary_departures,
+            secondary_departures: secondary_departures
           }
         } = config,
-        now,
-        fetch_departures_fn \\ fn params, opts ->
-          Departure.fetch(params, Keyword.put(opts, :include_scheduled_cancelled?, true))
-        end,
-        fetch_alerts_fn \\ &Alert.fetch_or_empty_list/1,
-        fetch_schedules_fn \\ &Screens.Schedules.Schedule.fetch/2,
-        fetch_routes_fn \\ &Screens.Routes.Route.fetch/1,
-        fetch_vehicles_fn \\ &Screens.Vehicles.Vehicle.by_route_and_direction/2
+        now
       ) do
-    [primary_instance, secondary_instance] =
-      Enum.map([primary_sections, secondary_sections], fn sections ->
-        sections
-        |> get_sections_data(fetch_departures_fn, fetch_alerts_fn, fetch_routes_fn, now)
-        |> sections_data_to_instance_fn(config, now, fetch_schedules_fn, fetch_vehicles_fn)
-      end)
+    primary_rds_sections = @rds.get(primary_departures, now)
 
-    # When the secondary instance would not be displaying anything useful, make it a copy of the
-    # primary instance.
-    secondary_instance =
-      if secondary_sections == [] or no_departures?(secondary_instance),
-        do: primary_instance,
-        else: secondary_instance
+    secondary_rds_sections = @rds.get(secondary_departures, now)
+
+    primary_departure_sections =
+      create_departure_sections(primary_rds_sections, primary_departures)
+
+    secondary_departure_sections =
+      if secondary_rds_sections == [] or Enum.all?(secondary_rds_sections, &(&1 == {:ok, []})) do
+        primary_departure_sections
+      else
+        create_departure_sections(secondary_rds_sections, secondary_departures)
+      end
+
+    all_sections_no_data =
+      Enum.all?(
+        primary_departure_sections ++ secondary_departure_sections,
+        &is_struct(&1, NoDataSection)
+      )
+
+    all_sections_service_ended =
+      Enum.all?(
+        primary_departure_sections ++ secondary_departure_sections,
+        &is_struct(&1, OvernightSection)
+      )
 
     primary_instances =
-      Enum.map(
-        [
-          :main_content_zero,
-          :main_content_one,
-          :main_content_reduced_zero,
-          :main_content_reduced_one
-        ],
-        &struct!(primary_instance, slot_names: [&1])
+      build_instances(
+        @primary_slot_names,
+        primary_departure_sections,
+        all_sections_no_data,
+        all_sections_service_ended,
+        config,
+        now
       )
 
     secondary_instances =
-      Enum.map(
-        [:main_content_two, :main_content_reduced_two],
-        &struct!(secondary_instance, slot_names: [&1])
+      build_instances(
+        @secondary_slot_names,
+        secondary_departure_sections,
+        all_sections_no_data,
+        all_sections_service_ended,
+        config,
+        now
       )
 
-    instances = primary_instances ++ secondary_instances
+    primary_instances ++ secondary_instances
+  end
+
+  defp create_departure_sections(rds_sections, %Departures{sections: departure_sections}) do
+    section_count = length(rds_sections)
+
+    Enum.zip(rds_sections, departure_sections)
+    |> Enum.map(fn {rds_section, %Section{bidirectional: bidirectional}} ->
+      map_to_departure_section(rds_section, bidirectional, section_count)
+    end)
+  end
+
+  @spec map_to_departure_section(RDS.section_t(), boolean(), number()) ::
+          DeparturesWidget.section()
+
+  defp map_to_departure_section(:error, _, _), do: %NoDataSection{}
+
+  defp map_to_departure_section({:ok, []}, _, _), do: %NoDataSection{}
+
+  defp map_to_departure_section({:ok, rds_list}, bidirectional, section_count) do
+    num_departures_per_section = div(@max_departures_per_rotation, section_count)
 
     cond do
-      # If every rotation is showing OvernightDepartures, we don't need to render any route pills.
-      Enum.all?(instances, &is_struct(&1, OvernightDepartures)) ->
-        Enum.map(instances, &%OvernightDepartures{&1 | routes: []})
-
-      # If every rotation consists entirely of NoDataSections, replace all with DeparturesNoData.
-      Enum.all?(instances, &no_departures?/1) ->
-        Enum.map(instances, fn %DeparturesWidget{screen: screen, slot_names: [slot_name]} ->
-          %DeparturesNoData{screen: screen, slot_name: slot_name}
-        end)
-
-      true ->
-        instances
-    end
-  end
-
-  defp no_departures?(%DeparturesWidget{sections: sections}),
-    do: Enum.all?(sections, &is_struct(&1, NoDataSection))
-
-  defp no_departures?(_widget_instance), do: false
-
-  @typep section_data ::
-           %{type: :no_data_section, route: Route.t()}
-           | %{
-               departures: [Departure.t()],
-               alert_informed_entities: [InformedEntity.t()],
-               stop_ids: [Stop.id()],
-               routes: [Route.t()],
-               params: Params.t()
-             }
-
-  @spec sections_data_to_instance_fn(
-          [section_data()],
-          Screen.t(),
-          DateTime.t(),
-          Schedule.fetch_with_date(),
-          Vehicle.by_route_and_direction()
-        ) :: DeparturesWidget.t() | OvernightDepartures.t()
-  defp sections_data_to_instance_fn(
-         sections_data,
-         config,
-         now,
-         fetch_schedules_fn,
-         fetch_vehicles_fn
-       ) do
-    is_only_section = match?([_], sections_data)
-
-    sections =
-      Enum.map(
-        sections_data,
-        &get_section_instance(&1, is_only_section, now, fetch_schedules_fn, fetch_vehicles_fn)
-      )
-
-    # NB: No slot names provided here (defaults to `[]`) as they will be filled in depending on
-    # whether the widget is placed in the primary or secondary slots.
-    if Enum.any?(sections) and Enum.all?(sections, &is_struct(&1, OvernightSection)) do
-      route_pills = get_route_pills_for_rotation(sections)
-      %OvernightDepartures{screen: config, routes: route_pills}
-    else
-      %DeparturesWidget{screen: config, sections: sections, now: now}
-    end
-  end
-
-  @spec get_section_instance(
-          section_data(),
-          boolean(),
-          DateTime.t(),
-          Schedule.fetch_with_date(),
-          Vehicle.by_route_and_direction()
-        ) :: DeparturesWidget.section()
-  defp get_section_instance(
-         %{type: :no_data_section, route: route},
-         _is_only_section,
-         _now,
-         _fetch_schedules_fn,
-         _fetch_vehicles_fn
-       ),
-       do: %NoDataSection{route: route}
-
-  defp get_section_instance(
-         %{
-           departures: departures,
-           alert_informed_entities: alert_informed_entities,
-           stop_ids: stop_ids,
-           routes: routes,
-           params: params
-         },
-         is_only_section,
-         now,
-         fetch_schedules_fn,
-         fetch_vehicles_fn
-       ) do
-    routes_with_live_departures =
-      departures
-      |> Enum.map(&{Departure.route(&1).id, Departure.direction_id(&1)})
-      |> Enum.uniq()
-
-    max_visible_departures = if is_only_section, do: 4, else: 2
-    # Check if there is any room for overnight rows before running the logic.
-
-    route_and_direction_schedules =
-      get_route_direction_schedules(params, routes, now, fetch_schedules_fn)
-
-    {section_contains_active_route, overnight_schedules_for_section} =
-      if length(departures) >= max_visible_departures do
-        {false, []}
-      else
-        get_overnight_schedules_for_section(
-          route_and_direction_schedules,
-          routes_with_live_departures,
-          params,
-          routes,
-          alert_informed_entities,
-          now,
-          fetch_vehicles_fn
-        )
-      end
-
-    in_overnight_period =
-      !section_contains_active_route and
-        overnight_schedules_for_section != [] and
-        overnight_time_period?(route_and_direction_schedules, now)
-
-    representative_route = hd(routes)
-
-    alert_headsign =
-      headsign_for_alert(alert_informed_entities, stop_ids)
-
-    headway_time_range = headway_time_range(stop_ids, routes, now)
-
-    cond do
-      # Normal departures mode
-      departures != [] ->
-        # Add overnight departures to the end.
-        # This allows overnight departures to appear as we start to run out of predictions to show.
-        visible_departures =
-          Enum.take(departures ++ overnight_schedules_for_section, max_visible_departures)
-
-        # DUPs don't support Layout or Header for now
-        %NormalSection{rows: visible_departures, layout: %Layout{}, header: %Header{}}
-
-      # If we have no departures, go through the other options
-
-      # In cases of temporary terminals, we want to show headways with directional headsigns
-      alert_headsign && headway_time_range && !in_overnight_period ->
-        %HeadwaySection{
-          route: get_section_route_from_entities(stop_ids, alert_informed_entities),
-          time_range: headway_time_range,
-          headsign:
-            if is_only_section do
-              alert_headsign
-            else
-              direction_name(representative_route, params.direction_id)
-            end
+      no_service?(rds_list) ->
+        %NoServiceSection{
+          routes:
+            rds_list
+            |> Enum.flat_map(fn %RDS{state: %NoService{routes: routes}} -> routes end)
+            |> Enum.uniq()
         }
 
-      # If we're not showing a headway for alerts, and all routes in the section are disabled, so no departures are expected.
-      # In this case, the alerts widget will display info on the closure, so we return an empty section.
-      section_routes_disabled?(routes, params.direction_id, alert_informed_entities) ->
-        %NormalSection{rows: departures, layout: %Layout{}, header: %Header{}}
-
-      # If no departures, no alerts, and not in the overnight period, we show Headways
-      headway_time_range &&
-          (alert_informed_entities == [] && !in_overnight_period) ->
-        %HeadwaySection{
-          route: representative_route.id,
-          time_range: headway_time_range,
-          headsign: direction_name(representative_route, params.direction_id)
+      service_ended?(rds_list) ->
+        %OvernightSection{
+          routes:
+            rds_list
+            |> Enum.map(fn %RDS{
+                             state: %ServiceEnded{
+                               last_scheduled_departure: last_scheduled_departure
+                             }
+                           } ->
+              Departure.route(last_scheduled_departure)
+            end)
+            |> Enum.uniq()
         }
 
-      # No remaining departures or active routes, but we do have routes with overnight schedules
-      # Show a takeover Overnight section for the given route types
-      in_overnight_period ->
-        %OvernightSection{routes: routes}
+      headways?(rds_list) ->
+        create_headway_section(rds_list)
 
-      # No departures to show and no headway mode
       true ->
-        %NoDataSection{route: representative_route}
+        %NormalSection{
+          rows:
+            create_and_sort_rows(rds_list)
+            |> maybe_make_bidirectional(bidirectional)
+            |> Enum.take(num_departures_per_section),
+          layout: %Layout{},
+          header: %Header{}
+        }
     end
   end
 
-  # Determines if the active alerts for a section apply to all routes that are enabled in the section
-  @spec section_routes_disabled?([Route.t()], 0 | 1 | :both, [InformedEntity.t()]) :: boolean()
-  defp section_routes_disabled?(routes, direction_id, alert_informed_entities) do
-    case alert_informed_entities do
-      [] ->
-        false
-
-      _ ->
-        # Normalize direction_id. Typically `nil` in Informed Entity represents both directions
-        direction_id =
-          case direction_id do
-            :both -> nil
-            _ -> direction_id
-          end
-
-        # For each route, verify if there is an associated Informed Entity
-        routes
-        |> Enum.map(& &1.id)
-        |> Enum.all?(fn route_id ->
-          Enum.any?(alert_informed_entities, fn entity ->
-            InformedEntity.present_alert_for_route?(entity, route_id, direction_id)
-          end)
-        end)
-    end
+  defp no_service?(rds_list) do
+    Enum.all?(rds_list, &is_struct(&1.state, NoService))
   end
 
-  @spec get_sections_data(
-          [Section.t()],
-          Departure.fetch(),
-          Alert.fetch(),
-          Route.fetch(),
-          now :: DateTime.t()
-        ) :: [section_data()]
-  defp get_sections_data(
-         sections,
-         fetch_departures_fn,
-         fetch_alerts_fn,
-         fetch_routes_fn,
-         now
-       ) do
-    sections
-    |> Task.async_stream(
-      &get_section_data(&1, fetch_departures_fn, fetch_alerts_fn, fetch_routes_fn, now),
-      on_timeout: :kill_task
-    )
-    |> Enum.map(fn
-      {:ok, data} ->
-        data
-
-      {:exit, reason} ->
-        Logger.error("get_section_data.exit reason=#{reason}")
-        raise "Failed to get section data"
-    end)
+  defp service_ended?(rds_list) do
+    Enum.all?(rds_list, &is_struct(&1.state, ServiceEnded))
   end
 
-  defp get_section_data(
-         %Section{query: %Query{params: %Params{stop_ids: stop_ids} = params}} = section,
-         fetch_departures_fn,
-         fetch_alerts_fn,
-         fetch_routes_fn,
-         now
-       ) do
-    routes = get_routes_serving_section(params, fetch_routes_fn)
-    # DUP sections will always show no more than one mode.
-    # For subway, each route will have its own section.
-    # If the stop is served by two different subway/light rail routes, route_ids must be populated for each section
-    # Otherwise, we only need the first route in the list of routes serving the stop.
-    primary_route_for_section = List.first(routes)
-
-    disabled_modes = Screens.Config.Cache.disabled_modes()
-
-    # If we know the predictions are unreliable, don't even bother fetching them.
-    if is_nil(primary_route_for_section) or primary_route_for_section.type in disabled_modes do
-      %{type: :no_data_section, route: primary_route_for_section}
-    else
-      section_departures =
-        case Widgets.Departures.fetch_section_departures(
-               section,
-               disabled_modes,
-               fetch_departures_fn
-             ) do
-          {:ok, departures} -> departures
-          :error -> []
-        end
-
-      alert_informed_entities = get_section_entities(params, fetch_alerts_fn, now)
-
-      %{
-        departures: section_departures,
-        alert_informed_entities: alert_informed_entities,
-        stop_ids: stop_ids,
-        routes: routes,
-        params: params
-      }
-    end
+  defp headways?(rds_list) do
+    Enum.all?(rds_list, &is_struct(&1.state, Headways))
   end
 
-  defp get_section_route_from_entities(
-         ["place-" <> _ = stop_id],
-         informed_entities
-       ) do
-    Enum.find_value(informed_entities, "", fn
-      %InformedEntity{route: route, stop: %{id: ^stop_id}} -> route
-      _ -> nil
-    end)
-  end
+  @spec create_headway_section([RDS.t()]) :: HeadwaySection.t()
 
-  defp get_section_route_from_entities(_, _), do: nil
-
-  defp get_section_entities(
-         %Params{
-           stop_ids: stop_ids,
-           route_ids: route_ids,
-           direction_id: direction_id
-         },
-         fetch_alerts_fn,
-         now
-       ) do
-    alert_fetch_params = [
-      direction_id: direction_id,
-      route_ids: route_ids,
-      stop_ids: stop_ids,
-      route_types: [:light_rail, :subway]
-    ]
-
-    # This section gets alert entities, which are used to decide whether we should be in headway mode or overnight mode
-    # Also used to check if no departures are expected for a section because of closures to all routes/directions
-    alert_fetch_params
-    |> fetch_alerts_fn.()
-    |> Enum.filter(fn
-      # Show a headway message only during shuttles and suspensions at temporary terminals.
-      # https://www.notion.so/mbta-downtown-crossing/Departures-Widget-Specification-20da46cd70a44192a568e49ea47e09ac?pvs=4#e43086abaadd465ea8072502d6980d8d
-      %Alert{effect: effect} = alert when effect in [:suspension, :shuttle] ->
-        Alert.happening_now?(alert, now)
-
-      _ ->
-        false
-    end)
-    # Condense all alerts into just a list of informed entities
-    # This will help us decide whether a headway in one direction is still useful
-    # if there are two alerts that could be in different directions.
-    |> Enum.reduce([], fn alert, acc -> acc ++ alert.informed_entities end)
-    |> Enum.uniq()
-  end
-
-  defp headsign_for_alert([], _), do: nil
-
-  defp headsign_for_alert(informed_entities, stop_ids) do
-    # Use all informed_entities from relevant alerts to decide whether there's
-    # a reason to go into headway mode for an alert
-    # For example, a NB suspension and SB shuttle from Aquarium shouldn't use headway mode
-    # but if all the WB branches are shuttling at Kenmore, there should be a headway for the alert
-    interpreted_alert = interpret_entities(informed_entities, stop_ids)
-
-    if temporary_terminal?(interpreted_alert) and
-         not (branch_station?(stop_ids) and branch_alert?(interpreted_alert)) do
-      interpreted_alert.headsign
-    else
-      nil
-    end
-  end
-
-  # NB: There aren't currently any DUPs at permanent terminals, so we assume all
-  # terminals are temporary. In the future, we'll need to check that the boundary
-  # isn't a normal terminal.
-  defp temporary_terminal?(%{region: :boundary}), do: true
-  defp temporary_terminal?(_), do: false
-
-  defp branch_station?(stop_ids) do
-    branch_stations = Application.get_env(:screens, :dup_headway_branch_stations)
-
-    case stop_ids do
-      [parent_station_id] -> parent_station_id in MapSet.new(branch_stations)
-      _ -> false
-    end
-  end
-
-  defp branch_alert?(%{headsign: headsign}) do
-    branch_terminals = Application.get_env(:screens, :dup_headway_branch_terminals)
-    headsign in MapSet.new(branch_terminals)
-  end
-
-  defp interpret_entities(entities, parent_stop_ids) do
-    informed_stop_ids = Enum.into(entities, MapSet.new(), & &1.stop.id)
-    parent_stop_id = List.first(parent_stop_ids)
-
-    {region, headsign} =
-      :screens
-      |> Application.get_env(:dup_alert_headsign_matchers)
-      |> Map.get(parent_stop_id, [])
-      |> Enum.find_value({:inside, nil}, fn
-        %{
-          informed: informed,
-          not_informed: not_informed,
-          headway_headsign: headsign
-        } ->
-          if alert_region_match?(
-               Util.to_set(informed),
-               Util.to_set(not_informed),
-               informed_stop_ids
-             ),
-             do: {:boundary, headsign},
-             else: false
-
-        _ ->
-          false
-      end)
-
-    %{
-      region: region,
-      headsign: headsign
+  # Bidirectional -> Use no headsign for the trains message
+  defp create_headway_section([
+         %RDS{state: %{route_id: route_id, direction_id: direction_id_one, range: range}}
+         | [%RDS{state: %{route_id: route_id, direction_id: direction_id_two, range: range}}]
+       ])
+       when direction_id_one != direction_id_two do
+    %HeadwaySection{
+      route: route_id,
+      time_range: range,
+      headsign: nil
     }
   end
 
-  defp alert_region_match?(informed, not_informed, informed_stop_ids) do
-    MapSet.subset?(informed, informed_stop_ids) and
-      MapSet.disjoint?(not_informed, informed_stop_ids)
-  end
-
-  defp overnight_time_period?(route_and_direction_schedules, now) do
-    {first_time, last_time} =
-      route_and_direction_schedules
-      |> Enum.reduce({nil, nil}, fn {_route_and_direction,
-                                     {first_schedule_today, last_schedule_today,
-                                      _first_schedule_tomorrow}},
-                                    {first_scheduled_time, last_scheduled_time} ->
-        first_time =
-          cond do
-            is_nil(first_scheduled_time) ->
-              first_schedule_today.departure_time
-
-            DateTime.compare(first_schedule_today.departure_time, first_scheduled_time) == :lt ->
-              first_schedule_today.departure_time
-
-            true ->
-              first_scheduled_time
-          end
-
-        last_time =
-          cond do
-            is_nil(last_scheduled_time) ->
-              last_schedule_today.departure_time
-
-            DateTime.compare(last_schedule_today.departure_time, last_scheduled_time) == :gt ->
-              last_schedule_today.departure_time
-
-            true ->
-              last_scheduled_time
-          end
-
-        {first_time, last_time}
-      end)
-
-    DateTime.compare(now, first_time) == :gt or
-      DateTime.compare(now, last_time) == :lt
-  end
-
-  # Return 'overnight' Departures from `get_overnight_schedules_for_section` to potentially show
-  # if one of the following is true for a given route_id/direction_id combo:
-  # 1. Service for the route is done for the day, so we may display the first departure of tomorrow.
-  # 2. Service for the route has not started for the day, so we may display the first departure of today.
-  # 3. Service for the route is done for the day and not scheduled tomorrow
-  #    (possible for CR/buses/routes with interruptions tomorrow), so return a Departure
-  #    with nil departure_time and arrival_time to be handled by the serializer.
-  #
-  # Returns a tuple where:
-  #  - First value is a boolean to track if there are any remaining scheduled trips for the section.
-  #  - Second value is a list of Departure Schedules to potentially show on screen given enough space.
-  defp get_overnight_schedules_for_section(
-         route_and_direction_schedules,
-         routes_with_live_departures,
-         stop_ids,
-         routes_serving_section,
-         alert_informed_entities,
-         now,
-         fetch_vehicles_fn
-       )
-
-  # No predictions AND no active alerts for the section
-  defp get_overnight_schedules_for_section(
-         route_and_direction_schedules,
-         routes_with_live_departures,
-         _params,
-         _routes,
-         [],
-         now,
-         _
+  # Use the headsign if the destinations have the same headsign,
+  # use the direction name if they have the same direction name,
+  # otherwise default to no headsign
+  defp create_headway_section(
+         [
+           %RDS{
+             headsign: headsign,
+             line: %Line{id: first_line_id},
+             state: %Headways{
+               route_id: route_id,
+               direction_name: direction_name,
+               direction_id: direction_id,
+               range: range
+             }
+           }
+           | _
+         ] = destinations
        ) do
-    # Get schedules for each route_id in config
-    overnight_schedules =
-      route_and_direction_schedules
-      |> Enum.reject(fn {route_and_direction, _schedule} ->
-        route_and_direction in routes_with_live_departures
-      end)
-      |> Enum.map(fn {{route_id, direction_id},
-                      {first_schedule_today, last_schedule_today, first_schedule_tomorrow}} ->
-        get_overnight_departure_for_route(
-          first_schedule_today,
-          last_schedule_today,
-          first_schedule_tomorrow,
-          route_id,
-          direction_id,
-          now
-        )
-      end)
+    %HeadwaySection{
+      route: route_id,
+      time_range: range,
+      headsign:
+        cond do
+          Enum.all?(destinations, fn %RDS{headsign: other_headsign} ->
+            headsign == other_headsign
+          end) ->
+            headsign
 
-    # We flag if any routes have service ongoing so the section doesn't show as fully overnight
-    is_sections_service_active = Enum.any?(overnight_schedules, &(&1 == :route_service_ongoing))
+          Enum.all?(
+            destinations,
+            fn %RDS{
+                 line: %Line{id: other_line_id},
+                 state: %Headways{direction_id: other_direction_id}
+               } ->
+              first_line_id == other_line_id and direction_id == other_direction_id
+            end
+          ) ->
+            direction_name
 
-    {is_sections_service_active,
-     overnight_schedules
-     # Routes not in overnight mode will be nil. Can ignore those.
-     |> Enum.reject(fn elem -> is_nil(elem) or elem == :route_service_ongoing end)
-     |> Enum.sort_by(fn %Departure{schedule: %Schedule{departure_time: dt}} -> dt end)}
+          true ->
+            nil
+        end
+    }
   end
 
-  defp get_overnight_schedules_for_section(
-         route_and_direction_schedules,
-         routes_with_live_departures,
-         params,
-         [%{type: type} | _] = routes,
-         alert_informed_entities,
-         now,
-         fetch_vehicles_fn
-       )
-       when type in [:subway, :light_rail] do
-    informed_route = get_section_route_from_entities(params.stop_ids, alert_informed_entities)
-    # If there are no vehicles operating on the route, assume we are overnight.
-    if not is_nil(informed_route) and fetch_vehicles_fn.(informed_route, nil) == [] do
-      get_overnight_schedules_for_section(
-        route_and_direction_schedules,
-        routes_with_live_departures,
-        params,
-        routes,
-        [],
-        now,
-        fetch_vehicles_fn
-      )
-    else
-      {false, []}
-    end
-  end
-
-  defp get_overnight_schedules_for_section(_, _, _, _, _, _, _), do: {false, []}
-
-  # Verifies we are meeting the timeframe conditions for overnight mode and generates the departure widget
-  defp get_overnight_departure_for_route(
-         first_schedule_today,
-         last_schedule_today,
-         _first_schedule_tomorrow,
-         route_id,
-         direction_id,
+  defp build_instances(
+         slot_names,
+         _departure_sections,
+         true = _all_section_no_data,
+         _all_section_service_ended,
+         config,
          _now
-       )
-       when is_nil(first_schedule_today) or is_nil(last_schedule_today) do
-    Report.warning("dup_overnight_no_first_last_schedule",
-      route_id: route_id,
-      direction_id: direction_id
-    )
-
-    nil
-  end
-
-  defp get_overnight_departure_for_route(
-         first_schedule_today,
-         last_schedule_today,
-         nil,
-         _route_id,
-         _direction_id,
-         now
        ) do
-    # If now is after today's last schedule and there are no schedules tomorrow,
-    # we still want a departure row without a time (will show a moon icon)
-    if DateTime.compare(now, last_schedule_today.departure_time) == :gt or
-         DateTime.compare(now, first_schedule_today.departure_time) == :lt do
-      # nil/nil acts as a flag for the serializer to produce an `overnight` departure time
-      %Departure{
-        schedule: %{last_schedule_today | departure_time: nil, arrival_time: nil}
-      }
-    else
-      # Return an atom so we can track that there is still a departure for this route during today's service.
-      :route_service_ongoing
-    end
+    Enum.map(slot_names, &%DeparturesNoData{screen: config, slot_name: &1})
   end
 
-  # If now is before any of today's schedules or after any of tomorrow's (should never happen but just in case)
-  # we do not display overnight mode.
-  defp get_overnight_departure_for_route(
-         first_schedule_today,
-         last_schedule_today,
-         first_schedule_tomorrow,
-         route_id,
-         direction_id,
+  defp build_instances(
+         slot_names,
+         _departure_sections,
+         _all_section_no_data,
+         true = _all_section_service_ended,
+         config,
+         _now
+       ) do
+    Enum.map(slot_names, &%OvernightDepartures{screen: config, slot_names: [&1]})
+  end
+
+  defp build_instances(
+         slot_names,
+         departure_sections,
+         _all_section_no_data,
+         _all_section_service_ended,
+         config,
          now
        ) do
     cond do
-      DateTime.compare(now, first_schedule_tomorrow.departure_time) == :gt ->
-        Report.warning("dup_overnight_after_first_schedule",
-          route_id: route_id,
-          direction_id: direction_id
+      Enum.all?(departure_sections, &is_struct(&1, NoServiceSection)) ->
+        Enum.map(
+          slot_names,
+          &%DeparturesNoService{
+            screen: config,
+            slot_name: &1,
+            routes:
+              departure_sections
+              |> Enum.flat_map(fn %NoServiceSection{routes: routes} -> routes end)
+              |> Enum.map(fn route -> Route.icon(route) end)
+              |> Enum.uniq()
+          }
         )
 
-        nil
-
-      # Before 4am and between the `departure_time` for today's last schedule and tomorrow's first schedule
-      DateTime.compare(now, last_schedule_today.departure_time) == :gt and
-          DateTime.compare(now, first_schedule_tomorrow.departure_time) == :lt ->
-        %Departure{schedule: first_schedule_tomorrow}
-
-      # After 4am but before the first scheduled trip of the day.
-      not is_nil(first_schedule_today) and
-          DateTime.compare(now, first_schedule_today.departure_time) == :lt ->
-        %Departure{schedule: first_schedule_today}
-
-      # Before 4am and before the last scheduled trip of the day.
-      # Return an atom so we can track that there is still a departure for this route during today's service.
-      DateTime.compare(now, last_schedule_today.departure_time) == :lt ->
-        :route_service_ongoing
+      Enum.all?(departure_sections, &is_struct(&1, OvernightSection)) ->
+        Enum.map(
+          slot_names,
+          &%OvernightDepartures{
+            screen: config,
+            slot_names: [&1],
+            routes:
+              departure_sections
+              |> Enum.flat_map(fn %OvernightSection{routes: routes} -> routes end)
+              |> Enum.map(fn route -> Route.icon(route) end)
+              |> Enum.uniq()
+          }
+        )
 
       true ->
-        nil
+        Enum.map(
+          slot_names,
+          &sections_to_departure_widget(&1, departure_sections, config, now)
+        )
     end
   end
 
-  defp get_today_tomorrow_schedules(
-         fetch_params,
-         fetch_schedules_fn,
-         now,
-         route_ids_serving_section
-       ) do
-    today =
-      case fetch_schedules_fn.(fetch_params, now) do
-        {:ok, schedules} when schedules != [] ->
-          Enum.filter(schedules, &(&1.route.id in route_ids_serving_section))
-
-        # fetch error or empty schedules
-        _ ->
-          []
-      end
-
-    tomorrow =
-      case fetch_schedules_fn.(fetch_params, now |> Util.service_date() |> Date.add(1)) do
-        {:ok, schedules} when schedules != [] ->
-          Enum.filter(schedules, &(&1.route.id in route_ids_serving_section))
-
-        # fetch error or empty schedules
-        _ ->
-          []
-      end
-
-    {today, tomorrow}
+  defp sections_to_departure_widget(slot_name, departure_sections, config, now) do
+    %DeparturesWidget{
+      screen: config,
+      sections: departure_sections,
+      slot_names: [slot_name],
+      now: now
+    }
   end
 
-  defp get_route_direction_schedules(params, routes, now, fetch_schedules_fn) do
-    {today_schedules, tomorrow_schedules} =
-      get_today_tomorrow_schedules(
-        params |> Map.from_struct() |> Map.put(:sort, "departure_time"),
-        fetch_schedules_fn,
-        now,
-        Enum.map(routes, & &1.id)
-      )
+  @spec create_and_sort_rows([RDS.t()]) :: [NormalSection.row()]
+  defp create_and_sort_rows(rds_list) do
+    grouped_rds =
+      Enum.group_by(rds_list, fn
+        %RDS{state: %ServiceEnded{}} -> :service_ended
+        %RDS{state: %Headways{}} -> :headways
+        _ -> :other
+      end)
 
-    today_schedules
-    |> Enum.map(&{&1.route.id, &1.direction_id})
-    |> Enum.uniq()
-    |> Enum.map(fn {route_id, direction_id} ->
-      # This variable will be used when now is after the start-of-service time.
-      first_schedule_today =
-        Enum.find(
-          today_schedules,
-          &(&1.route.id == route_id and &1.direction_id == direction_id)
-        )
+    service_ended_rds = Map.get(grouped_rds, :service_ended, [])
+    headway_rds = Map.get(grouped_rds, :headways, [])
+    rds = Map.get(grouped_rds, :other, [])
 
-      last_schedule_today =
-        Enum.find(
-          Enum.reverse(today_schedules),
-          &(&1.route.id == route_id and &1.direction_id == direction_id)
-        )
+    sorted_departures_from_rds(rds) ++
+      headways_from_rds(rds ++ service_ended_rds, headway_rds) ++
+      sorted_departures_from_rds(service_ended_rds, true)
+  end
 
-      first_schedule_tomorrow =
-        Enum.find(
-          tomorrow_schedules,
-          &(&1.route.id == route_id and &1.direction_id == direction_id)
-        )
+  @spec maybe_make_bidirectional([Departure.t()], boolean()) :: [Departure.t()]
+  defp maybe_make_bidirectional([], _), do: []
+  defp maybe_make_bidirectional(departures, false), do: departures
 
-      {{route_id, direction_id},
-       {first_schedule_today, last_schedule_today, first_schedule_tomorrow}}
+  defp maybe_make_bidirectional([first | rest], true) do
+    first_direction = departure_direction_id(first)
+
+    opposite? =
+      Enum.find(rest, Enum.at(rest, 0), &(departure_direction_id(&1) == 1 - first_direction))
+
+    Enum.reject([first, opposite?], &is_nil/1)
+  end
+
+  @spec sorted_departures_from_rds([RDS.t()], boolean()) :: [Departure.t()]
+  defp sorted_departures_from_rds(rds, reverse \\ false) do
+    sort_order =
+      if reverse do
+        :desc
+      else
+        :asc
+      end
+
+    rds
+    |> Enum.flat_map(&departure_rows_from_state(&1))
+    |> Enum.sort_by(
+      &departure_time(&1),
+      {
+        sort_order,
+        DateTime
+      }
+    )
+  end
+
+  @spec headways_from_rds([RDS.t()], [RDS.t()]) :: [HeadwayRow.t()]
+  defp headways_from_rds(
+         rds,
+         headway_rds
+       ) do
+    # If there are other similar line/direction_id destinations that are in one of the other states,
+    # disregard the headway for that particular destination
+    lines_in_other_states =
+      rds
+      |> Enum.flat_map(&extract_line_direction_pairs/1)
+      |> MapSet.new()
+
+    headway_rds
+    |> Enum.reject(fn %RDS{
+                        line: %Line{id: line_id},
+                        state: %Headways{direction_id: direction_id}
+                      } ->
+      MapSet.member?(lines_in_other_states, {line_id, direction_id})
+    end)
+    |> Enum.group_by(fn %RDS{
+                          line: line,
+                          state: %Headways{direction_id: direction_id}
+                        } ->
+      {line, direction_id}
+    end)
+    |> Enum.flat_map(fn {{line, direction_id}, rds_list} ->
+      %RDS{
+        headsign: headsign,
+        state: %Headways{
+          departure_id: departure_id,
+          direction_name: direction_name,
+          range: range
+        }
+      } = hd(rds_list)
+
+      # If there are multiple headways with the same line but different headsigns,
+      # combine them and use the direction name
+      displayed_headsign = if length(rds_list) == 1, do: headsign, else: direction_name
+
+      [
+        %HeadwayRow{
+          id: departure_id,
+          line: line,
+          direction_id: direction_id,
+          range: range,
+          headsign: displayed_headsign
+        }
+      ]
     end)
   end
 
-  defp get_routes_serving_section(
-         %Params{route_ids: route_ids, route_type: route_type, stop_ids: stop_ids},
-         fetch_routes_fn
-       ) do
-    params =
-      %{stop_ids: stop_ids}
-      |> then(&if route_ids == [], do: &1, else: Map.put(&1, :ids, route_ids))
-      |> then(&if is_nil(route_type), do: &1, else: Map.put(&1, :route_types, [route_type]))
+  @spec departure_rows_from_state(RDS.t()) ::
+          [Departure.t()] | [{Departure.t(), NormalSection.special_trip_type()}]
+  defp departure_rows_from_state(%RDS{state: %Countdowns{departures: departures}}), do: departures
 
-    case fetch_routes_fn.(params) do
-      {:ok, routes} -> routes
-      :error -> []
+  defp departure_rows_from_state(%RDS{
+         state: %FirstTrip{first_scheduled_departure: first_scheduled_departure}
+       }) do
+    [{first_scheduled_departure, :first_trip}]
+  end
+
+  defp departure_rows_from_state(%RDS{
+         state: %ServiceEnded{last_scheduled_departure: last_scheduled_departure}
+       }) do
+    [{last_scheduled_departure, :last_trip}]
+  end
+
+  defp departure_rows_from_state(%RDS{state: %NoService{}}), do: []
+
+  @spec departure_time(Departure.t()) :: DateTime.t()
+  defp departure_time(%Departure{} = departure), do: Departure.time(departure)
+
+  defp departure_time({first_scheduled_departure, :first_trip}),
+    do: Departure.time(first_scheduled_departure)
+
+  defp departure_time({last_scheduled_departure, :last_trip}),
+    do: Departure.time(last_scheduled_departure)
+
+  defp departure_direction_id(%Departure{} = departure), do: Departure.direction_id(departure)
+
+  defp departure_direction_id({first_scheduled_departure, :first_trip}),
+    do: Departure.direction_id(first_scheduled_departure)
+
+  defp departure_direction_id({last_scheduled_departure, :last_trip}),
+    do: Departure.direction_id(last_scheduled_departure)
+
+  defp departure_direction_id(%HeadwayRow{direction_id: direction_id}), do: direction_id
+
+  defp extract_line_direction_pairs(%RDS{line: %Line{id: line_id}, state: state}) do
+    case state do
+      %NoService{direction_id: nil} ->
+        []
+
+      %NoService{direction_id: direction_id} ->
+        [{line_id, direction_id}]
+
+      %Countdowns{departures: [first_departure | _]} ->
+        [{line_id, Departure.direction_id(first_departure)}]
+
+      %Countdowns{departures: [], direction_id: direction_id} ->
+        [{line_id, direction_id}]
+
+      %FirstTrip{first_scheduled_departure: departure} ->
+        [{line_id, Departure.direction_id(departure)}]
+
+      %ServiceEnded{last_scheduled_departure: departure} ->
+        [{line_id, Departure.direction_id(departure)}]
     end
-  end
-
-  defp get_route_pills_for_rotation(sections) do
-    sections
-    |> Enum.flat_map(fn %OvernightSection{routes: routes} -> Enum.map(routes, &Route.icon/1) end)
-    |> Enum.uniq()
-  end
-
-  defp headway_time_range(stop_ids, routes, now) do
-    all_headways =
-      for stop_id <- stop_ids, %{id: route_id} <- routes do
-        @headways.get_with_route(stop_id, route_id, now)
-      end
-
-    case Enum.uniq(all_headways) do
-      [{lo, hi}] -> {lo, hi}
-      _ -> nil
-    end
-  end
-
-  defp direction_name(_, :both), do: nil
-
-  defp direction_name(route, direction_id) do
-    route
-    |> Route.normalized_direction_names()
-    |> Enum.at(direction_id, nil)
   end
 end
