--- lib/screens/v2/candidate_generator/dup_new.ex
@@ -1,30 +1,0 @@
-defmodule Screens.V2.CandidateGenerator.DupNew do
-  @moduledoc false
-
-  alias Screens.V2.CandidateGenerator
-  alias Screens.V2.CandidateGenerator.Dup.Alerts, as: AlertsGenerator
-  alias Screens.V2.CandidateGenerator.Widgets
-
-  alias __MODULE__.Departures
-
-  @behaviour CandidateGenerator
-
-  @instance_generators [
-    &Widgets.Header.instances/2,
-    &Departures.instances/2,
-    &Widgets.Evergreen.evergreen_content_instances/2,
-    &AlertsGenerator.alert_instances/2,
-    &Widgets.EmergencyTakeover.emergency_takeover_instances/2
-  ]
-
-  @impl CandidateGenerator
-  defdelegate screen_template(screen), to: Screens.V2.CandidateGenerator.Dup
-
-  @impl CandidateGenerator
-  def candidate_instances(config, now \\ DateTime.utc_now()) do
-    CandidateGenerator.async_stream(@instance_generators, & &1.(config, now), timeout: 20_000)
-  end
-
-  @impl CandidateGenerator
-  def audio_only_instances(_widgets, _config), do: []
-end
