--- lib/screens/v2/candidate_generator/dup_new/departures.ex
@@ -1,468 +1,0 @@
-defmodule Screens.V2.CandidateGenerator.DupNew.Departures do
-  @moduledoc false
-
-  import Screens.Inject
-
-  alias Screens.Lines.Line
-  alias Screens.Routes.Route
-
-  alias Screens.V2.Departure
-  alias Screens.V2.RDS
-  alias Screens.V2.RDS.{Countdowns, FirstTrip, Headways, NoService, ServiceEnded}
-
-  alias Screens.V2.WidgetInstance.Departures, as: DeparturesWidget
-
-  alias Screens.V2.WidgetInstance.Departures.{
-    HeadwayRow,
-    HeadwaySection,
-    NoDataSection,
-    NormalSection,
-    NoServiceSection,
-    OvernightSection
-  }
-
-  alias Screens.V2.WidgetInstance.{DeparturesNoData, DeparturesNoService, OvernightDepartures}
-
-  alias ScreensConfig.Departures
-  alias ScreensConfig.Departures.{Header, Layout, Section}
-  alias ScreensConfig.Screen
-  alias ScreensConfig.Screen.Dup
-
-  @type widget :: DeparturesNoData.t() | DeparturesWidget.t() | OvernightDepartures.t()
-  @rds injected(RDS)
-
-  @max_departures_per_rotation 4
-
-  @primary_slot_names [
-    :main_content_zero,
-    :main_content_one,
-    :main_content_reduced_zero,
-    :main_content_reduced_one
-  ]
-  @secondary_slot_names [:main_content_two, :main_content_reduced_two]
-
-  @spec instances(Screen.t(), DateTime.t()) :: [widget()]
-  def instances(
-        %Screen{
-          app_params: %Dup{
-            primary_departures: primary_departures,
-            secondary_departures: secondary_departures
-          }
-        } = config,
-        now
-      ) do
-    primary_rds_sections = @rds.get(primary_departures, now)
-
-    secondary_rds_sections = @rds.get(secondary_departures, now)
-
-    primary_departure_sections =
-      create_departure_sections(primary_rds_sections, primary_departures)
-
-    secondary_departure_sections =
-      if secondary_rds_sections == [] or Enum.all?(secondary_rds_sections, &(&1 == {:ok, []})) do
-        primary_departure_sections
-      else
-        create_departure_sections(secondary_rds_sections, secondary_departures)
-      end
-
-    all_sections_no_data =
-      Enum.all?(
-        primary_departure_sections ++ secondary_departure_sections,
-        &is_struct(&1, NoDataSection)
-      )
-
-    all_sections_service_ended =
-      Enum.all?(
-        primary_departure_sections ++ secondary_departure_sections,
-        &is_struct(&1, OvernightSection)
-      )
-
-    primary_instances =
-      build_instances(
-        @primary_slot_names,
-        primary_departure_sections,
-        all_sections_no_data,
-        all_sections_service_ended,
-        config,
-        now
-      )
-
-    secondary_instances =
-      build_instances(
-        @secondary_slot_names,
-        secondary_departure_sections,
-        all_sections_no_data,
-        all_sections_service_ended,
-        config,
-        now
-      )
-
-    primary_instances ++ secondary_instances
-  end
-
-  defp create_departure_sections(rds_sections, %Departures{sections: departure_sections}) do
-    section_count = length(rds_sections)
-
-    Enum.zip(rds_sections, departure_sections)
-    |> Enum.map(fn {rds_section, %Section{bidirectional: bidirectional}} ->
-      map_to_departure_section(rds_section, bidirectional, section_count)
-    end)
-  end
-
-  @spec map_to_departure_section(RDS.section_t(), boolean(), number()) ::
-          DeparturesWidget.section()
-
-  defp map_to_departure_section(:error, _, _), do: %NoDataSection{}
-
-  defp map_to_departure_section({:ok, []}, _, _), do: %NoDataSection{}
-
-  defp map_to_departure_section({:ok, rds_list}, bidirectional, section_count) do
-    num_departures_per_section = div(@max_departures_per_rotation, section_count)
-
-    cond do
-      no_service?(rds_list) ->
-        %NoServiceSection{
-          routes:
-            rds_list
-            |> Enum.flat_map(fn %RDS{state: %NoService{routes: routes}} -> routes end)
-            |> Enum.uniq()
-        }
-
-      service_ended?(rds_list) ->
-        %OvernightSection{
-          routes:
-            rds_list
-            |> Enum.map(fn %RDS{
-                             state: %ServiceEnded{
-                               last_scheduled_departure: last_scheduled_departure
-                             }
-                           } ->
-              Departure.route(last_scheduled_departure)
-            end)
-            |> Enum.uniq()
-        }
-
-      headways?(rds_list) ->
-        create_headway_section(rds_list)
-
-      true ->
-        %NormalSection{
-          rows:
-            create_and_sort_rows(rds_list)
-            |> maybe_make_bidirectional(bidirectional)
-            |> Enum.take(num_departures_per_section),
-          layout: %Layout{},
-          header: %Header{}
-        }
-    end
-  end
-
-  defp no_service?(rds_list) do
-    Enum.all?(rds_list, &is_struct(&1.state, NoService))
-  end
-
-  defp service_ended?(rds_list) do
-    Enum.all?(rds_list, &is_struct(&1.state, ServiceEnded))
-  end
-
-  defp headways?(rds_list) do
-    Enum.all?(rds_list, &is_struct(&1.state, Headways))
-  end
-
-  @spec create_headway_section([RDS.t()]) :: HeadwaySection.t()
-
-  # Bidirectional -> Use no headsign for the trains message
-  defp create_headway_section([
-         %RDS{state: %{route_id: route_id, direction_id: direction_id_one, range: range}}
-         | [%RDS{state: %{route_id: route_id, direction_id: direction_id_two, range: range}}]
-       ])
-       when direction_id_one != direction_id_two do
-    %HeadwaySection{
-      route: route_id,
-      time_range: range,
-      headsign: nil
-    }
-  end
-
-  # Use the headsign if the destinations have the same headsign,
-  # use the direction name if they have the same direction name,
-  # otherwise default to no headsign
-  defp create_headway_section(
-         [
-           %RDS{
-             headsign: headsign,
-             line: %Line{id: first_line_id},
-             state: %Headways{
-               route_id: route_id,
-               direction_name: direction_name,
-               direction_id: direction_id,
-               range: range
-             }
-           }
-           | _
-         ] = destinations
-       ) do
-    %HeadwaySection{
-      route: route_id,
-      time_range: range,
-      headsign:
-        cond do
-          Enum.all?(destinations, fn %RDS{headsign: other_headsign} ->
-            headsign == other_headsign
-          end) ->
-            headsign
-
-          Enum.all?(
-            destinations,
-            fn %RDS{
-                 line: %Line{id: other_line_id},
-                 state: %Headways{direction_id: other_direction_id}
-               } ->
-              first_line_id == other_line_id and direction_id == other_direction_id
-            end
-          ) ->
-            direction_name
-
-          true ->
-            nil
-        end
-    }
-  end
-
-  defp build_instances(
-         slot_names,
-         _departure_sections,
-         true = _all_section_no_data,
-         _all_section_service_ended,
-         config,
-         _now
-       ) do
-    Enum.map(slot_names, &%DeparturesNoData{screen: config, slot_name: &1})
-  end
-
-  defp build_instances(
-         slot_names,
-         _departure_sections,
-         _all_section_no_data,
-         true = _all_section_service_ended,
-         config,
-         _now
-       ) do
-    Enum.map(slot_names, &%OvernightDepartures{screen: config, slot_names: [&1]})
-  end
-
-  defp build_instances(
-         slot_names,
-         departure_sections,
-         _all_section_no_data,
-         _all_section_service_ended,
-         config,
-         now
-       ) do
-    cond do
-      Enum.all?(departure_sections, &is_struct(&1, NoServiceSection)) ->
-        Enum.map(
-          slot_names,
-          &%DeparturesNoService{
-            screen: config,
-            slot_name: &1,
-            routes:
-              departure_sections
-              |> Enum.flat_map(fn %NoServiceSection{routes: routes} -> routes end)
-              |> Enum.map(fn route -> Route.icon(route) end)
-              |> Enum.uniq()
-          }
-        )
-
-      Enum.all?(departure_sections, &is_struct(&1, OvernightSection)) ->
-        Enum.map(
-          slot_names,
-          &%OvernightDepartures{
-            screen: config,
-            slot_names: [&1],
-            routes:
-              departure_sections
-              |> Enum.flat_map(fn %OvernightSection{routes: routes} -> routes end)
-              |> Enum.map(fn route -> Route.icon(route) end)
-              |> Enum.uniq()
-          }
-        )
-
-      true ->
-        Enum.map(
-          slot_names,
-          &sections_to_departure_widget(&1, departure_sections, config, now)
-        )
-    end
-  end
-
-  defp sections_to_departure_widget(slot_name, departure_sections, config, now) do
-    %DeparturesWidget{
-      screen: config,
-      sections: departure_sections,
-      slot_names: [slot_name],
-      now: now
-    }
-  end
-
-  @spec create_and_sort_rows([RDS.t()]) :: [NormalSection.row()]
-  defp create_and_sort_rows(rds_list) do
-    grouped_rds =
-      Enum.group_by(rds_list, fn
-        %RDS{state: %ServiceEnded{}} -> :service_ended
-        %RDS{state: %Headways{}} -> :headways
-        _ -> :other
-      end)
-
-    service_ended_rds = Map.get(grouped_rds, :service_ended, [])
-    headway_rds = Map.get(grouped_rds, :headways, [])
-    rds = Map.get(grouped_rds, :other, [])
-
-    sorted_departures_from_rds(rds) ++
-      headways_from_rds(rds ++ service_ended_rds, headway_rds) ++
-      sorted_departures_from_rds(service_ended_rds, true)
-  end
-
-  @spec maybe_make_bidirectional([Departure.t()], boolean()) :: [Departure.t()]
-  defp maybe_make_bidirectional([], _), do: []
-  defp maybe_make_bidirectional(departures, false), do: departures
-
-  defp maybe_make_bidirectional([first | rest], true) do
-    first_direction = departure_direction_id(first)
-
-    opposite? =
-      Enum.find(rest, Enum.at(rest, 0), &(departure_direction_id(&1) == 1 - first_direction))
-
-    Enum.reject([first, opposite?], &is_nil/1)
-  end
-
-  @spec sorted_departures_from_rds([RDS.t()], boolean()) :: [Departure.t()]
-  defp sorted_departures_from_rds(rds, reverse \\ false) do
-    sort_order =
-      if reverse do
-        :desc
-      else
-        :asc
-      end
-
-    rds
-    |> Enum.flat_map(&departure_rows_from_state(&1))
-    |> Enum.sort_by(
-      &departure_time(&1),
-      {
-        sort_order,
-        DateTime
-      }
-    )
-  end
-
-  @spec headways_from_rds([RDS.t()], [RDS.t()]) :: [HeadwayRow.t()]
-  defp headways_from_rds(
-         rds,
-         headway_rds
-       ) do
-    # If there are other similar line/direction_id destinations that are in one of the other states,
-    # disregard the headway for that particular destination
-    lines_in_other_states =
-      rds
-      |> Enum.flat_map(&extract_line_direction_pairs/1)
-      |> MapSet.new()
-
-    headway_rds
-    |> Enum.reject(fn %RDS{
-                        line: %Line{id: line_id},
-                        state: %Headways{direction_id: direction_id}
-                      } ->
-      MapSet.member?(lines_in_other_states, {line_id, direction_id})
-    end)
-    |> Enum.group_by(fn %RDS{
-                          line: line,
-                          state: %Headways{direction_id: direction_id}
-                        } ->
-      {line, direction_id}
-    end)
-    |> Enum.flat_map(fn {{line, direction_id}, rds_list} ->
-      %RDS{
-        headsign: headsign,
-        state: %Headways{
-          departure_id: departure_id,
-          direction_name: direction_name,
-          range: range
-        }
-      } = hd(rds_list)
-
-      # If there are multiple headways with the same line but different headsigns,
-      # combine them and use the direction name
-      displayed_headsign = if length(rds_list) == 1, do: headsign, else: direction_name
-
-      [
-        %HeadwayRow{
-          id: departure_id,
-          line: line,
-          direction_id: direction_id,
-          range: range,
-          headsign: displayed_headsign
-        }
-      ]
-    end)
-  end
-
-  @spec departure_rows_from_state(RDS.t()) ::
-          [Departure.t()] | [{Departure.t(), NormalSection.special_trip_type()}]
-  defp departure_rows_from_state(%RDS{state: %Countdowns{departures: departures}}), do: departures
-
-  defp departure_rows_from_state(%RDS{
-         state: %FirstTrip{first_scheduled_departure: first_scheduled_departure}
-       }) do
-    [{first_scheduled_departure, :first_trip}]
-  end
-
-  defp departure_rows_from_state(%RDS{
-         state: %ServiceEnded{last_scheduled_departure: last_scheduled_departure}
-       }) do
-    [{last_scheduled_departure, :last_trip}]
-  end
-
-  defp departure_rows_from_state(%RDS{state: %NoService{}}), do: []
-
-  @spec departure_time(Departure.t()) :: DateTime.t()
-  defp departure_time(%Departure{} = departure), do: Departure.time(departure)
-
-  defp departure_time({first_scheduled_departure, :first_trip}),
-    do: Departure.time(first_scheduled_departure)
-
-  defp departure_time({last_scheduled_departure, :last_trip}),
-    do: Departure.time(last_scheduled_departure)
-
-  defp departure_direction_id(%Departure{} = departure), do: Departure.direction_id(departure)
-
-  defp departure_direction_id({first_scheduled_departure, :first_trip}),
-    do: Departure.direction_id(first_scheduled_departure)
-
-  defp departure_direction_id({last_scheduled_departure, :last_trip}),
-    do: Departure.direction_id(last_scheduled_departure)
-
-  defp departure_direction_id(%HeadwayRow{direction_id: direction_id}), do: direction_id
-
-  defp extract_line_direction_pairs(%RDS{line: %Line{id: line_id}, state: state}) do
-    case state do
-      %NoService{direction_id: nil} ->
-        []
-
-      %NoService{direction_id: direction_id} ->
-        [{line_id, direction_id}]
-
-      %Countdowns{departures: [first_departure | _]} ->
-        [{line_id, Departure.direction_id(first_departure)}]
-
-      %Countdowns{departures: [], direction_id: direction_id} ->
-        [{line_id, direction_id}]
-
-      %FirstTrip{first_scheduled_departure: departure} ->
-        [{line_id, Departure.direction_id(departure)}]
-
-      %ServiceEnded{last_scheduled_departure: departure} ->
-        [{line_id, Departure.direction_id(departure)}]
-    end
-  end
-end
