+++ test/screens/headways_test.exs
@@ -20,7 +20,6 @@ defmodule Screens.HeadwaysTest do
     @sl_one_terminal_a "17092"
     @sl_two_drydock_ave "31259"
     @sl_three_airport "7096"
-    @congress_st_at_wtc "17096"
 
     test "returns nil for a stop with no defined headways" do
       assert Headways.get("nonexistent", local_dt()) == nil
@@ -44,23 +43,6 @@ defmodule Screens.HeadwaysTest do
       assert Headways.get(@blue_trunk, local_dt(1, 17)) == {5, 7}
       assert Headways.get(@blue_trunk, local_dt(6)) == {10, 14}
       assert Headways.get(@blue_trunk, local_dt(7)) == {12, 16}
-    end
-  end
-
-  describe "get_with_route/3" do
-    test "returns the correct value for a combination of parent station and route for Rail" do
-      assert Headways.get_with_route("place-aport", "Blue", local_dt()) == {9, 13}
-      assert Headways.get_with_route("place-north", "Green-D", local_dt()) == {7, 13}
-      assert Headways.get_with_route("place-north", "Orange", local_dt()) == {9, 11}
-    end
-
-    test "returns the correct value for a combination of station id and route for the Silver Line" do
-      assert Headways.get_with_route("place-chels", "743", local_dt()) == {7, 9}
-      assert Headways.get_with_route("place-crtst", "746", local_dt()) == {1, 3}
-      assert Headways.get_with_route("place-crtst", "743", local_dt()) == {1, 3}
-      assert Headways.get_with_route(@congress_st_at_wtc, "742", local_dt()) == {1, 3}
-      assert Headways.get_with_route(@congress_st_at_wtc, "743", local_dt()) == {1, 3}
-      assert Headways.get_with_route("place-crtst", "751", local_dt()) == nil
     end
   end
 end
